pyproject.toml
[project]
name = "ostrovok-hackaton"
version = "1.0.0"
description = "A secret agent service for assessing hotel quality."
authors = [
    {name = "macht1212", email = "nasimov.alexander@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.117.1,<0.118.0",
    "sqlalchemy>=2.0.43,<3.0.0",
    "alembic>=1.16.5,<2.0.0",
    "pydantic[email]>=2.11.9,<3.0.0",
    "pydantic-settings>=2.11.0,<3.0.0",
    "uvicorn>=0.37.0,<0.38.0",
    "asyncpg>=0.30.0,<0.31.0",
    "minio>=7.2.17,<8.0.0",
    "python-multipart (>=0.0.20,<0.0.21)",
    "passlib[bcrypt]>=1.7.4",
    "PyJWT>=2.10.1",
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
black = "^25.9.0"
flake8 = "^7.3.0"
flake8-black = "^0.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
    | \.venv
    | migrations
    | build
    | dist
    | __pycache__
    | tests
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "--verbose",
    "--color=yes",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=xml",
    "--cov-report=html"
]
